educred-ml/requirements.txt
Flask==2.2.5
pytesseract==0.3.10
Pillow==9.3.0
rapidfuzz
openai
pdf2image
